Phase 3 bucket 2b: F821 part 2 — 17 files, genuine undefined-name fixes (missing imports, py2 builtins, typos)
- AmberFormat: py2 sys.exc_type/exc_value -> sys.exc_info(); add missing sys/traceback imports
- FilterClouds: math.sqrt -> sqrt (module already does 'from math import sqrt')
- universal/Util: long(x) -> int(x); FitClass: bare getValue -> self.getValue (subclass dispatch)
- Typos: MoleculeAlign mappping->mapping; workflow/Util tgzFile/tgzFileName, zipFile/zipFileName;
  TabularFormat lineSep -> self.lineSep (class attr); exception -> Exception
- Missing imports: getSoftware (FactorisedParams), copy (DataConvertLib),
  vectorsAdd/vectorsSubtract (ViewChemCompVarFrame), pnt2ppm (NoeRelaxation),
  getPeakDimPpm (PseudoResonances), estimateAssignmentTolerances/findMatchingPeakDimShifts/
  areResonancesBound (AssignmentAdvanced)
- GUI/local: IsdFrame 'except Exception' -> 'as msg'; DataLocationFrame __main__ local import
  of DataLocationPopup (avoids cycle); MultiSelectionList gains alwaysOnTop kwarg already
  passed to startBasePopup; CCPNReader constraint_index -> constraint.serial (file idiom, 15+ uses)

Gates: smoke 1643/0/83, compileall 0, pytest 15/14/10 (10 = data-gated, unchanged)

+++ ccpnmr2.5/python/memops/universal/Util.py
@@ -333,7 +333,7 @@ def returnLong(x, default=0.0, verbose=1):
     """
 
     try:
-        x = long(x)
+        x = int(x)
     except:
         if verbose:
             print("Error converting '" + str(x) + "' to long: set to %s" % str(default))
